Revert "refactor(63): split view into an optional `window` group; make nt_window opt-out"
This reverts commit 5f7a2c6b7dc75b685132c6912fa2e32f15fd3491.
