+++ engine/devapi/nt_devapi_registry.c
@@ -45,9 +45,6 @@ nt_result_t nt_devapi_init(void) {
 #ifdef NT_DEVAPI_REGISTER_core
     nt_devapi_register_core();
 #endif
-#ifdef NT_DEVAPI_REGISTER_window
-    nt_devapi_register_window();
-#endif
 
     /* Discovery is always-on when devapi is built. */
     nt_devapi_register_discovery();
